requirements.txt
absl-py >= 0.7.0
pandas >= 0.25.0
python-dateutil >= 2.8.0
numpy < 2.0,  >= 1.16.0
scikit-learn >= 0.19.0
scipy < 2.0, >= 1.3.1
tensorflow < 3.0, >= 2.0.0
transformers <3.0, >= 2.0
tensorflow-datasets < 2.0, >= 1.3